% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.ezpzr.R
\name{cor.ezpzr}
\alias{cor.ezpzr}
\title{Compute pairwise (weighted) correlations between all columns of a dataset with (bootstrapped) confidence intervals and return matrices, including significance stars.}
\usage{
cor.ezpzr(
  data,
  method = "pearson",
  cis = TRUE,
  weight = NULL,
  nboot = 1000,
  correction = c("none", "bonferroni", "fdr"),
  alpha = 0.05
)
}
\arguments{
\item{data}{Data frame of numeric variables}

\item{method}{Correlation method ("pearson", "spearman", "kendall")}

\item{cis}{Logical, include confidence intervals?}

\item{weight}{Optional column name for weights}

\item{nboot}{Number of bootstrap samples}

\item{correction}{Correction to be applied c("none","bonferroni","fdr")}

\item{alpha}{Alpha level (default .05)}
}
\value{
A list with correlation matrices, p-values, stars, and CIs
}
\description{
A helper function for psychologists to compute correlations,
p-values, significance stars, and confidence intervals (standard and bootstrap),
optionally weighted.
}
